split ing version specific functions
